Make vulnerabilityManagedFilters function common
