+++ plugins/main/public/components/overview/vulnerabilities/dashboards/inventory/inventory.tsx
@@ -61,7 +61,7 @@ import {
   VULNERABILITIES_INVENTORY_DASHBOARD_ID,
   VULNERABILITIES_INVENTORY_AGENT_DASHBOARD_ID,
 } from '../../../../../../common/constants';
-import { vulnerabilityManagedFilters } from '../overview/dashboard';
+import { vulnerabilityManagedFilters } from '../../common/vulnerability-managed-filters';
 import RestoreStateColumnsButton from '../../../../common/wazuh-discover/components/restore-state-columns';
 import managedFilters from './managed-filters';
 import DashboardRenderer from '../../../../common/dashboards/dashboard-renderer/dashboard-renderer';
